requirements.txt
browser-use
langchain-ollama
python-dotenv